+++ doc/scripting/encryption-rot13.qml
@@ -1,9 +1,17 @@
 import QtQml 2.0
 
 /**
- * This script will encrypt and decrypt notes with PGP
+ * This script will "encrypt" and "decrypt" notes with ROT13
  */
 QtObject {
+    /**
+     * This is called when the script is loaded by QOwnNotes
+     */
+    function init() {
+        // disable the password dialog
+        script.encryptionDisablePassword();
+    }
+
     /**
      * This function is called when text has to be encrypted or decrypted
      * 
@@ -13,19 +21,11 @@ QtObject {
      * @return the exncrypted or decrypted text
      */
     function encryptionHook(text, password, decrypt) {
-        script.log("password");
-        script.log(password);
-        script.log("before");
-        script.log(text);
-        
         // do the rot13 transformation
         text = text.replace(/[a-zA-Z]/g, function(c){
             return String.fromCharCode((c <= "Z" ? 90 : 122) >= (c = c.charCodeAt(0) + 13) ? c : c - 26);
         });
         
-        script.log("after");
-        script.log(text);
-
         return text;
     }
 }